admin_panel/app/requirements.txt
asgiref==3.7.2
Django==4.2.5
django-cors-headers==4.3.1
django-debug-toolbar==4.2.0
django-extensions==3.2.3
django-split-settings==1.2.0
flake8==6.1.0
mccabe==0.7.0
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pyflakes==3.1.0
python-dotenv==1.0.0
sqlparse==0.4.4
style==1.1.0
update==0.0.1
uWSGI==2.0.23
